requirements.txt
cffi>=1.0.0
Flask>=0.10.1
Flask-Bootstrap>=3.3.2.1
Flask-Cors>=2.1.2
pycparser>=2.10
pyalsaaudio>=0.8
gevent>=1.0.1
pyspotify-connect>=0.1.0